split API class to APIClient
